requirements.txt
dj-database-url==0.5.0
Django==2.1.6
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
